Finish scraping

+++ grapchcool_updater/graphcool_deleter.js
@@ -9,7 +9,7 @@ if (process.argv[2] !== 'all') {
   genders = [process.argv[2]];
 }
 
-let categories = ["abrigos_chaquetas", "camisas_blusas", "camisetas", "faldas", "monos", "pantalones_cortos", "pantalones_largos", "punto", "sudaderas_jerseis", "tops_bodies", "vestidos"];
+let categories = ["abrigos_chaquetas", "camisas_blusas", "camisetas", "faldas", "monos", "pantalones_cortos", "pantalones_largos", "sudaderas_jerseis", "tops_bodies", "vestidos"];
 if (process.argv[3] !== 'all') {
     categories = [process.argv[3]]
 }
@@ -41,18 +41,20 @@ genders.forEach((gender) => {
       		}
     	  }
         `).then((res) => {
-          client.mutate(`
-            {
-              deleteProduct(
-                id: "${res.allProducts[0].id}"
-              ) {
-                id
+          if (res.allProducts.length) {
+            client.mutate(`
+              {
+                deleteProduct(
+                  id: "${res.allProducts[0].id}"
+                ) {
+                  id
+                }
               }
-            }
-          `)
-          .catch((error) => {
-            throw error;
-          });
+            `)
+            .catch((error) => {
+              throw error;
+            });
+          }
         })
         .catch((error) => {
           console.log(error);
